+++ app/assets/js/views/docsearch.ts
