+++ app/assets/js/views/index.ts
@@ -2,7 +2,7 @@ import type { Views } from "@icelab/defo";
 
 import { breakpointFilter } from "~/utils/breakpoints";
 import { lazyLoadView } from "~/utils/lazy-load";
-import { docsearchViewFn } from "./docsearch/docsearch";
+import { docsearchViewFn } from "./docsearch";
 import { sizeToVarViewFn } from "./size-to-var";
 import { tocScrollViewFn } from "./toc-scroll";
 import { toggleClassViewFn } from "./toggle-class";
